requirements.txt
PySide6>=6.5
matplotlib>=3.7
numpy>=1.25
pyqtgraph>=0.13
pywebview>=3.8
